+++ example/lib/presentation/src/theme/theme.dart
@@ -1,4 +1,5 @@
 export 'src/ac_calendar_header_theme_data.dart';
+export 'src/ac_calendar_theme.dart';
 export 'src/ac_calendar_theme_data.dart';
 export 'src/ac_day_theme_data.dart';
 export 'src/ac_week_theme_data.dart';